npm3
npm3 organizes the node_modules folder very differently than npm2.
Using a shrinkwrap.json file generated on a Mac with the new structure
was causing install errors in Windows and Unix. Best fix for OS
compatibility seems to be to ditch the shrinkwrap.json file and instead
specify exact dependency versions in package.json. Some but not as much
control as we previously had with a shrink-wrap file. C'est la vie!