case sensitivity
Better support for Linux default case sensitive volumes and macOS and Windows case insensitive volumes. Case detection is OS independent so any volume with any casing style should work.

If you have a volume that forces file names to all uppercase or lowercase, I'd love to hear about your setup. At the time of this commit I did not have an upper or lowercase forced volume to test with.

This commit also includes various improvements that are not related to case sensitivity. Opportunities that were too tempting not to do while working on the main case sensitivity goal. ^_^